modif deletePost

+++ controllers/post.js
@@ -43,23 +43,31 @@ exports.modifyPost = (req, res, next) => {
 };
 
 exports.deletePost = (req, res, next) => {
-  const id = req.params.id
+  const id = req.params.id;
+  const token = req.headers.authorization.split(' ')[1];//On extrait le token de la requête
+  const decodedToken = jwt.verify(token, process.env.JWT_SECRET_KEY);//On décrypte le token grâce à la clé secrète
+  const userId = decodedToken.userId;//On récupère l'userId du token décrypté
+  const isAdmin = decodedToken.isAdmin;//On récupère l'userId du token décrypté
   Post.findOne({ where: { id: id } })
-      .then(post => {
+    .then(post => {
+      if(post.idUser == userId || isAdmin == true) {
         if (post.image !== null){
           const fileName = post.image.split('/images/')[1]
           fs.unlink(`images/${fileName}`, (err => {//On supprime l'ancienne image
             if (err) console.log(err);
             else {
-                console.log("Image supprimée: " + fileName);
+              console.log("Image supprimée: " + fileName);
             }
           }))
         }
         post.destroy({ where: { id: id } })
-            .then(() => res.status(200).json({  message: 'post supprimé !' }))
-            .catch(error => res.status(400).json({ error }));
-      })
-      .catch(error => res.status(500).json({ error }));
+          .then(() => res.status(200).json({  message: 'post supprimé !' }))
+          .catch(error => res.status(400).json({ error }));
+      }else {
+        return res.status(401).json({ error: "vous n'avez pas l'autorisation nécessaire !" });
+      }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.likePost = (req, res, next) => {
